generic-connector/requirements.txt
flask
pymongo
psycopg2-binary
mysql-connector-python